pyproject.toml
[tool.poetry]
name = "resume-analyzer"
version = "0.1.0"
description = ""
authors = ["Subash Basnet <subash.basnet0123@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
flashtext = "^2.7"
nltk = "^3.9.1"
python-docx = "^1.1.2"
beautifulsoup4 = "^4.12.3"
contractions = "^0.1.73"
numpy = "1.26.4"
poetry = "^1.8.4"
spacy = "3.7.4"
pandas = "^2.2.3"
scikit-learn = "1.2.2"
scipy = "1.10.1"
sentence-transformers = "^3.3.1"
gensim = "^4.3.3"
flair = "^0.14.0"
pdfplumber = "^0.11.4"
python-levenshtein = "^0.26.1"
fuzzywuzzy = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
